Refactored 'go to word' and 'go to start of word' commands to use utility function.
